llama-alfred-agent/pyproject.toml
[project]
name = "llama-alfred-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "datasets>=4.0.0",
    "huggingface-hub>=0.34.3",
    "llama-index-embeddings-huggingface>=0.6.0",
    "llama-index-llms-huggingface-api>=0.6.0",
    "llama-index-retrievers-bm25>=0.6.1",
    "python-dotenv>=1.1.1",
]
